requirements.txt
streamlit==1.28.0
pandas==2.0.3
databricks-sql-connector==3.0.0
plotly==5.17.0
altair==5.1.0
scikit-learn>=1.3.0
python-dateutil==2.8.2
requests==2.31.0
scipy>=1.9.0
streamlit-aggrid==1.1.7
python-dotenv==1.0.0 